+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The `pycldf` package adheres to [Semantic Versioning](http://semver.org/spec/v2.
 
 ## [Unreleased]
 
+- Support somewhat efficient data access in CLDFMarkdownText.
 - Fix bug whereby CLDF Markdown links to rows in custom tables were erased by
   FilenameToComponent.
 - Fix bug whereby CLDF Markdown links to metadata were not recognized as such.